20220623 1431

+++ Homework_04/app/src/main/java/com/moviesearch/datasource/database/FilmDao.kt
@@ -22,6 +22,6 @@ interface FilmDao {
     fun getLiked(idKp: Int): Long
 
     @Query("SELECT * FROM list_films WHERE list_films.page_id = :pageId")
-    fun getPageFilms(pageId: Int): MutableList<Film>
+    fun getPageFilms(pageId: Long): MutableList<Film>
 }
